Push operator images right before the next ones.
Unfortunately we need a separate function because the tag names are different
(.0 rather than .x) and not consistent (some have a "v" prefix).

Also they need to go first because the synchronization in other test jobs only
looks at the non-operator images.  This could in principle be changed, but the
pushes are surprisingly fast so we do not waste much time.