Get rid of the GO_VERSION operator build arg. (#2495)
Hardcode the go version in the dockerfile, rather than require it to be passed
as a build argument.

This may seem counter-productive at first, but there is an important detail
outside of this repository. Namely operator/Dockerile is also referenced
from the OpenShift CI configuration, which currently also needs to hardcode
the go version, since the value of the build arg there needs to be provided
literally (there does not seem to be a way to extract it from
EXPECTED_GO_VERSION).

So rather than have the Go version be defined in two separete repositories
(here and in openshift/release), we put it in two places in this repository,
and add CI- and build-time checks to make sure they are in sync, as well as
cross-references to remember to update them simultaneously.