Make `output_lint_test` use development_build flavor unconditionally
Also add missing `opensource` flavor to test.
It's not great to have to add missing flavor manually but I wasn't
able to come up how to make slice of all flavors available in
tests but not available in the main code.
We don't need it in the main code to prevent ourselves from
accidentally using it instead of `GetVisibleImageFlavorNames`.