Add operator architectures explicitly in Makefile
This follows up on https://github.com/stackrox/stackrox/pull/3163
That PR introduces some code which is for downstream to enable
multiple architectures there. The issue is the code is never called
upstream and so we're risking it gets rusty or may even get removed
as unnecessary.

Having a code that always works is better than having conditional
code that works rarely.