Pull k8s events in case of test failure as yaml file
The problem with `kubectl get events -o wide` is that the
`LAST SEEN` and the `FIRST SEEN` columns have _relative_ timestamps
like `4m11s`. This is very inconvenient in investigating CI failures
because one needs to know the instant when `kubectl get events` was
ran and do the math.

I suggest that we dump events in yaml format which preserves the
original info and absolute timestamps like `2022-10-05T10:31:32Z`.