Implement closing lifecycle for PLOP
PLOP objects should be distinguished to "active" and "closed" based on
the ClosedTimestamp value. If it's set to not null value, it means the
event we receive notifies about the endpoint beint closed. In that case
instead of creating a new record, update the existing one to be closed.

Such closed events are going to be ingored from the API.