Implement pruning for PLOP
There are chances to have PLOP objects stored in the db without properly
setup reference to a Process Indicator. They're not getting ignored to
not lose the data and for introspection purposes, but constantly
aggregating them is not possible. Thus extend pruning to clean up such
objects following the regular pruning rules (most importantly
orphanWindow). The implementation follow the established suboptimal
approach and has to be reworked in the future.