Move port and protocol fields out of serialized
There are two cases when we would like to be able to operate with port
and protocol fields on the database side:

* Doing queries like "which processes listen on port X"

* Adding where clauses into the existing queries to minimize amount of
data passed between the db and the application

To enable this, move out those two fields out of opaque serialized
column into corresponding separated columns. This would allow us to be
more flexible, without needing to perform any kind of schema and data
migrations in the future to support the cases mentioned before.