Drop the old Helm charts versioning code
I was looking at the code with both "old 3.y.z" and "new 4.y.z"
branches and thought that I need to write more comments to justify
the existence of the "old" code.
At the same time, we're probably not going to go back in time to
produce 3.y.z charts from 4.y.z code so I decided it can be killed.
Simply went ahead and killed it.