add test failing render if ImageTag isn't set
This is test should fail in this commit. The scenario where
MetaValues return empty ImageTag could happen if the Cluster
defined in the DB is using a SHA reference. Which is possible
on Operator-based installations.

ImageTag isn't used to render TLS secrets, therefore it shouldn't result
in RenderSensorTLSSecretsOnly returning an error.