Fix ownership of the current directory in CI
so that git can further work with code cloned by `actions.checkout`
without complaining about dubious ownership:

fatal: detected dubious ownership in repository at '/__w/stackrox/stackrox'

In our case, the default current directory is owned by
uid=1001,gid=123 but the current user is 0:0 (root).

`git` does not like it when the repo directory is owned by a
different user and refuses to work with it until the directory is
marked as safe or the ownership is fixed.

The interesting thing is that `actions.checkout@v3` does by default
mark the directory as safe. `set-safe-directory` is `true` by
default. However, that does not help.
Changing the ownership as done here helps. Therefore,
`set-safe-directory` is redundant and can be set to `false`.

Unfortunately we can't create a reusable action for `chown` in the
repo because the action is not visible before the repo is checked
out, but we need to fix things before checking out.