Revert "Raise alerts via syslog using CEF signature field generated from policy ID"
MS Sentinel documents this as "String or integer that serves as a
unique identifier per event type".

On second thoughts, "Alert" is a fine value to use for this field. The
name of the policy that was violated will appear in the CEF Name field
which Sentinel maps to its Activity field.

This reverts commit b4b716f1991dc994d8bd05528be19284d3171899.