add failing test: multiple tests starting servers
This is the crux of the problem this PR aims to solve.
It's not possible to start two servers in two separate tests, becuase
the server never closes until the program finishes. We are bound to the
lifecycle of Go runtime, and the server should never stop unless the
process stops.

This makes writing unit and integration tests in Sensor much harder,
because you can't have two tests that start a gRPC server in the same
file or package. They must be ran in separate Go runtimes in order to
be able to register the same server in the same port.

Test should fail with:

panic: Failed to instantiate endpoint config of kind TLS-enabled multiplexed HTTP/gRPC: listen tcp :8080: bind: address already in use