review: change API to make Stop() a blocking call
This makes Stop() synchronous and removes the need for two signals
to control the state of gRPC server.

However, it introduced the need to use a mutex in the loop make sure
that two goroutines won't call Stop() simultaneously.

This requires adding a TryLock implementation in our own sync package.