ROX-17992: Try bearer auth as fallback in jira notifier
Currently only basic auth is tried. If a user provides an oauth
token or JWT in the password field, then auth will fail.

With this change, a user can provide the token in the "password or api
token" field (with a dummy username value), and bearer auth will succeed
after a failed basic auth attempt.