ROX-18310: Add separate CI job specifically for Compliance e2e testing
Previously, we ran the compliance end-to-end testing with the UI
end-to-end testing. While this is good, and we should continue to do it
since compliance can impact UI functionality, we can separate the
compliance testing out into its own CI job.

This gives us the flexibility to invoke the compliance end-to-end
testing on backend-specific changes, instead of only UI related changes.