Revert Add ability to set `spec.skips` in CSV via `patch-csv.py`
This reverts commit d4faa492033bc47584357dc0509638eefde5ad0c.
Note that it's not a plain revert, I had to resolve a few conflicts
so that the newer code remains unaffected.

Originally, the ability to set `spec.skips` was added in
https://github.com/stackrox/stackrox/pull/4165, as part of ROX-13946
but remained unused both upstream and downstream.

Later on, static skips were introduced in
https://github.com/stackrox/stackrox/pull/6830 / ROX-18208.

I think, having both static and dynamic skips with the latter ones
unused is confusing and burdening, therefore I suggest removing this
ability of dynamic skips.