Set force rebuild flag to true
The one is piped via the `init` task to `init.results.build`, see
https://github.com/redhat-appstudio/build-definitions/blob/f8f6cf41b41997b0aa407ec61a130892beb09a3c/task/init/0.1/init.yaml#L56-L59

Also, the flag is describe there as
> Rebuild the image if exists

I don't see any reason for us to skip builds if we started a
pipeline. This feels like an optimization that could lead to
confusing results.

I suggest that we start with forced builds and see if we have a need
to introduce skipping.

Most of times we'll have commits which would change the output image
tag (under the current scheme) and so these will be rebuilt anyways.