[central] Remove useless and racy step.
After deleting a pod, nothing guarantees that the status.availableReplicas
value seen by the assert reflects the newly created pod. It might simply be
stale if the replication controller did not get round to doing its thing yet.

In fact, this whole step 50 is superfluous because step 40 already makes sure
the pod is re-created.