[upgrade] Reuse assert files in upgrade step 40.
Unfortunately the check for volume in 40-errors.yaml would never work as the
author intended (because of poor list support in kuttl).

Luckily, the central test step 91 already contains files which list complete
volume and volume mount lists, so we reuse these instead. However for that to
be possible we need to disable monitoring on the Central resource. Tracking
cleanup of this in ROX-20649.