ROX-18313: Introduce compliance v2 e2e tests
The v1 compliance operator testing in
`tests/compliance_operator_tests.go` relies on the ability to install
mocked data from the Compliance Operator. This helped ensure the data
the tests were asserting against didn't change, especially since the
data is ultimately coming from the [Compliance
Operator](https://github.com/ComplianceAsCode/compliance-operator) and
it's [content](https://github.com/ComplianceAsCode/content), which can
change across Compliance Operator releases.

However, the new v2 compliance work needs a test runner that can install
the Compliance Operator on a test cluster, and invoke a separate set of
compliance tests that exercise the integration with the Compliance
Operator through ACS APIs.

This commit attempts to do that by using a separate test runner
dedicated to testing v2 compliance functionality. The v1 compliance
functionality is still important, and will remain in place until it's
deprecated and removed, but it will be included in all e2e integration
runs like it is today.

This commit also adds a `README.md` to the `tests/complianceoperator`
directory so that it's clear that the CRDs and resources there are
intended for testing v1 compliance.