Merge/drop allowNone=true tests.
For secured cluster chart, this test currently does not provide any value over
the top-level expectations in the context of "with no image pull secret
creation"/"works with allowNone=true", given that it does not tolerate errors.

For central, its stronger (saOnlyRefersTo-based) counterpart still makes sense
because of the different secret arrangement ("stackrox" is both a default and
the primary secret there, unlike here). However it can be merged with another
similar test without any loss of coverage.