Refactor "with image pull secret creation" tests and merge in another one.
The useExisting feature is already tested in the "no secret creation" test. For
the "no secret creation" case, we do not need another standalone test, so merge
it in as a sub-test (with a sligtly different test data, consistent with the
other top-level test).

This is made possible by specifying the username value in the common root.