ROX-21409: Detect CO installs during integration tests
For CI purposes, we'll always need to install the Compliance Operator in
the cluster before the integration tests run, but locally that might not
be the case.

When developers have local clusters they're running integration tests
against, they might have CO installed already from a previous
integration test run.

This commit makes it easier for contributors to rerun complinace
integration tests by detecting if the CO CSV is present, and using it
instead of attempting to reinstall it.