ROX-18091: Implement rescan API
After a scan configuration exists, users should be able to rerun the
scan schedules on demand. This commit adds a new v2 API for invoking
that behavior and wires it up to send messages to the appropriate
sensors so they can annotate the Compliance Operator objects necessary
to rescan the cluster.