Create a custom, helm-free stackrox installer
This is the start of a pure-go installer.  Instead of Helm, the
kubernetes client-go library is used by constructing kube API structs
instead of reading in YAML templates.

This commit represents an MVP of sorts.  The installer will apply all
the resources to create a functional Central with basic auth configured.

An off-the-cuff roadmap:

* Add in viper/cobra for CLI configurability
* Refactor to be reused by operator
* Deploy scanner
* Deploy secured cluster components
* Accept all the parameters in helm chart
* HostPath volumes for local development