ROX-18375: Translate ebpf to core_bpf (#9356)
Introduce a new values translation in the operator to switch from ebpf collection method to core_bpf. It means at upgrade of an existing secured cluster, configured to use ebpf, it will be modified to use core_bpf.

The translation is happening on every sync, which means by itself it will prevent any attempts to override a secured cluster CR to temporary use ebpf. To still be able to use ebpf if needed, a new boolean configuration option is introduced spec.perNode.Collector.forceCollection. If the flag is set, no translation from ebpf to core_bpf would be applied, and ebpf value could be used.

The use case for the new option is temporary workarounds and troubleshooting, thus it doesn't have to be advertised in common scenarios in the documentation.