ROX-18124: Document eventual removal for permission workaround
This commit just leaves a reference to the story that tracks the work to
add permissions to the installation path, which should allow us to
remove this workaround in the future. This commit also removes the some
labels on the role and binding so they don't interfere with any of the
install tools.