ROX-21848: Fix compliance e2e tests to pass backend validation
The compliance e2e tests are failing with traces like the following:

    compliance_operator_v2_test.go:119:
        	Error Trace:	/go/src/github.com/stackrox/stackrox/tests/compliance_operator_v2_test.go:119
        	Error:      	Received unexpected error:
        	            	scan result not found
        	Test:       	TestComplianceV2CreateGetScanConfigurations
    --- FAIL: TestComplianceV2CreateGetScanConfigurations (39.17s)

This is because the scan configuration name contains an underscore
instead of a dash and fails the backend validation.

This commit updates the tests to use a dash instead of an underscore,
but there should probably be a follow up to determine why the assertions
aren't preventing the test from failing immediately after the scan
config is created.