ROX-21579: Add database column for scan types
Compliance scans have a type, either `Platform` or `Node`, which
correlates to the profile they were derived from. This type is useful in
determining which profile a result came from because we can map the
result to the scan, and the scan to the profile, but only if we have the
scan type.

This commit adds a new column in the V2 scan table to store the scan
type, which is already being forwarded along from sensor when new scans
are created by the compliance operator on secured clusters.