Allow inline tasks in pipelines
Otherwise we get errors like

```
Pipeline task 'fetch-sql-definitions' does not contain a bundle reference
```

with a suggested resolution

```
Check for the existence of a task bundle. This rule will fail if the task is not called from a bundle. To exclude this rule add "attestation_task_bundle.tasks_defined_in_bundle" to the `exclude` section of the policy configuration.
```