Add labels required by Enterprise Contract
These were not problematic before because we inherited them from
base containers.
In case of `main`, we don't inherit anything from `scratch` and so
EC fails without `distribution-scope`, `release` and `vendor`.

`release` is supposed to be an auto-incremented build number for a
specific version. We don't have that, so I added a todo.

`vendor` is easy.

`distribution-scope` was not very straightforward.
EC said:

  The required "distribution-scope" label is missing. Label description: Scope of intended distribution of the image. (private/authoritative-source-only/restricted/public).

I pulled UBI8 and RHEL8 containers and found it to be `public` in
both.
It is also `public` for RHACS containers from pretty much the
beginning of downstream builds.

Searching Slack did not return much fruitful besides this message
and the next one
https://redhat-internal.slack.com/archives/C04PZ7H0VA8/p1705593365314769?thread_ts=1705493582.556159&cid=C04PZ7H0VA8

Then I found this repo with maintainers from Red Hat
https://github.com/projectatomic/ContainerApplicationGenericLabels?tab=readme-ov-file#overview

It describes `public` as:

  No redistribution limits beyond licenses

And also mentions this w.r.t. `distribution-scope`:

  Allows a user to define the intended scope of distribution. This addresses the end-user case of internal builds vs. public content and the use case of a vendor like Red Hat that provides content streams under subscription agreements - which is different from the license(s) of the image content.

Therefore, I conclude it should be `public`.