Fix error assertions in compliance manager unit tests
The compliance manager uses table driven tests, and one of the
attributes of the test cases is an error string. When set, the handler
checks that the error returned from the function under test matches the
error string.

However, the assertion is using `Require().Error()`, which doesn't
assert the error strings are an exact match. This gives the impression that
the error string we're setting in the test is actually being asserted,
but it isn't. For example, if you modify any of the expected error
strings, the tests still pass.

This commit updates the tests to use `Require().ErrorContains()`, which
asserts the error contains a substring that we're defining as the
expected error string.