ROX-22610: Use smaller profiles for compliance e2e testing
We're noticing timeouts with the existing compliance e2e tests where the
compliance scan doesn't finish within a five minute window.

While we're investigating why this is taking longer than normal, let's
update the test to use a smaller profile the runs faster, especially
since the ACS functionality doesn't depend on the contents of the
profile.