Set memory requests==limits for buildah to address failing build
https://console.redhat.com/preview/application-pipeline/workspaces/rh-acs/applications/acs/pipelineruns/main-on-pull-request-zxdpp
failed with the known symptom during UI build:
```
[build] @stackrox/platform-app: The build failed because the process exited too early. This probably means the system ran out of memory or someone called `kill -9` on the process.
```

This is one rare case when the build failed, which in the absence of
other changes in my PR means that's something about the
infrastructure and not the code.

My initial reaction was to switch from buildah-6gb to buildah-8gb but
then I looked at definitions and found that buildah-6gb sets memory
limit=6gb and request=4gb.
https://github.com/redhat-appstudio/build-definitions/blob/main/task/buildah-6gb/0.1/patch.yaml

My hunch is that we saw a situation when we were given 4gb, tried to
go over it but failed due to other pods on the node.

Instead of trying to patch `buildah-6gb`, I switch to vanilla
`buildah` and override both requests and limits to the same value.

I hope it should help. We should see whether it's going to fail the
same way over time.