Refactor generation of image list.
Apart from checking image availability on Quay, this will soon be used
to provide a list of images to pre-fetch on a cluster.

Rather than pass around just names and then tweak tags and names
elsewhere, produce a canonical list just in one place.

This way is slightly more verbose, but simpler and more reusable.