Simplify exporting `CGO_ENABLED`
Doing export with `?=` assignment is functionally equivalent to the
previous version that did `ifeq`.

Except for the following edge case.

Previously, if `CGO_ENABLED` was set to empty string, the logic
redefined it to `1`.
After this change, if `CGO_ENABLED` is set to empty string (i.e.
explicitly set to that), it will stay so.

Go behavior when `CGO_ENABLED` environment variable is empty string
seems to use what `go env CGO_ENABLED` tells which for me is `1`.
Playing with `CGO_ENABLED=<value> go env CGO_ENABLED` only `0` value
renders `0` on the output.

I think, we should be fine in this edge case.