Releax restriction on profile name duplication
Let's relax the restriction on having duplicate profile with differnet version, previousely we made sure no same profile to be run at same time, even they have differernt name, ex. ocp-cis, and ocp-cis-1-5 is not allowed to run at the same time. However this is not required by Compliance Operator, as along as the same name profile only used once. So we can run ocp-cis with ocp-cis-1-5 and ocp-cis-1-5, but we can
 not reference ocp-cis in two scan.