Prevent port 0 from being used as a server
Collector sometimes doesn't get the port number of some sockets, this
may be due to it missing some syscall or the kernel simply not filling in
the data. This last case most commonly happens with connectionless UDP
messaging, where the kernel leaves the side that starts the
communication empty.

In any case, we shouldn't be showing port 0 in the network graph, so
this patch will make it so port 0 takes the least precedence, so it will
only show up if both sides of the connection are 0.