ci: Enable config-controller unit tests
I had originally excluded all config-controller unit tests, but now that
we have a few that need to be run in CI, I am including the
`config-controller` directory and am commenting out the tests that still
don't work.  The commented tests have a TODO to ensure we follow up on
this.