ROX-25828: Make Pruner more configurable
Currently the only thing that could be tuned for the pruning process is
whether it's enabled. The way how pruner works right now might be
causing high memory utilization, due to queries doing large amount of
reads. This could be adjusted via frequence and the orphan cut-off line.
Introduce those two configurations to help with troubleshooting.