ROX-25492: Deploy SecurityPolicy CRD alongside ACS operator
For operator-based installs of ACS, there is only one point in time
where CRDs can be installed: during install/ugprade of the ACS operator.
To put it another way: the SecurityPolicy CRD cannot be installed when a
Central CR is reconciled because the ACS operator itself does not have
permission to create CRDs.

This approach takes two steps:

1. Copy the CRD from config-controller over to the operator bundle
2. Add a reference to the SecurityPolicy CRD in the ACS operator CSV

An alternative approach was considered where the operator-sdk CLI was
"tricked" into generating the SecurityPolicy CRD, but this required
copying over the go structs for the SecurityPolicy CRD, which seemed
like much more work overall.