ROX-25828: Make Pruner more configurable (#12780)
Currently the only thing that could be tuned for the pruning process is whether
it's enabled. The way how pruner works right now might be causing high memory
utilization, due to queries doing large amount of reads. This could be adjusted
via frequency and the orphan cut-off line. Introduce those two configurations
to help with troubleshooting.