Remove tabs from postgresql conf
It turns out tabs in the configuration are causing an annoying
inconvenience. The yaml renderer output for files with tabs ends up
being a mess of all lines merged together with the new line character in
between. This makes manual configuration harder, and since the tabs are
not essential, get rid of them.