ROX-24986: Store discovered external entities in Central (#11746)
Unmatched entities are forwarded by Sensor to Central as
'discovered' external entities. They are the src or dst of the
reported network-flows, and include a CIDR describing
them. When such an entity is received by Central, it is
added in the (existing) network_entities table and the 
corresponding network-flow points to it.