Removing tests from config-controller dir
These are the tests created by the initial kubebuilder scaffolding.

There was a TODO to re-enable these tests, but now that there is an e2e
test in the parent `tests` folder, I don't think these tests are going
to be used.

The policy-as-code e2e tests in the parent tests folder are more easily
included in already-existing test suites, and we can always re-add this
later if we really feel the e2e testing is lacking.