Add index on process_indicators.signal_time
Process indicator pruning worker struggles if the table grows too large.
One of the reasons is that limiting query results only to the old
records is not supported by any index:

    signal_time < now() AT time zone 'utc' - INTERVAL '$1 MINUTES'

Add the index process_indicators_signal_time_idx, which together with
the configuration option for orphan window allow us to specify how much
data to prune for process indicators.