Exclude CRDs from helm customize test
CRDs are cluster-scoped resources, so applying labels and annotations
could be confusing if more than one ACS installation is on a cluster.

Also, applying the templating is tricky because the CRD content is
generated, thus there would have to be a tool to automatically patch the
CRD with the templating lines every time the CRD is regenerated.