fix: Increase number of connection retries from config-controller (#13026)
The config-controller tries to establish a gRPC connection to Central, but if Central isn't ready yet, it retries until it can connect. If it eventually can't connect after 10 tries, it will die. The thinking is that the restarting of the container could resolve some transient or platform issue.

The problem is that in CI, Central is often not ready after 10 attempts, and a restarting pod fails some jobs. Therefore increase the number of retries 10x to make sure it should never reasonably happen in CI.