fix: Rely on liveness probe to restart config-controller container
After increasing the retry count in #13026, I was still observing
container restarts.  It turns out the liveness probe was killing the
container.  Since it seems like the liveness probe is a better fit for
this job (killing the container if it's not ready), I decided to
simplify the initialization code with the assumption that the liveness
probe would restart the container after some time if it wasn't ready.

To calculate how long the liveness probe will wait, use these values
from the config:

initialDelaySeconds + (periodSeconds * failureThreshold)

These values before this PR:

15s + (20s * 3) = 65s  # 3 is the default failureThreshold

These values after this PR:

15s + (20s * 50) = 1015s  # Roughly 17 minutes

A failureThreshold of 50 was chosen because:

* Long enough to avoid CI failures solely due to unexpected container restarts
* 50 is a nice, round number
* 17 minutes is generally not a long time to go without reconciliation for these resources

A couple more notes:

* Stopped using the retry function and went for a simple `for` loop with a 5 second sleep.  This loop will never stop unless the process is killed (see above) or the connection is successful.
* Took the client struct initialization out of the for loop.  It turns out that no network connections are attempted during struct construction.