Simplify source image copying
It happens so that all our scanner* images are retagged in the same
repo without copying and so the conditional which checks if the
repo would change is the only one that works.

In some sense it's unfair that we don't check whether the source
image exists because we don't fail, but, hey, is checking the source
image responsibility of the retagging pipeline?

This closes https://issues.redhat.com/browse/ROX-26721