ROX-26630: Ensure policy CRs can't apply to default policies (#13091)
If a policy is renamed to something that already exists, reconcilation fails. If the CR is then deleted, the finalizer is never removed because the config-controller can never delete the policy in Central. The right behavior is to check to see if the policy CR is currently marked as accepted before trying to delete the policy from Central.

This does potentially leave an orphaned policy in Central, but there's not much we can do about that after the CR changed the name. Perhaps if we tried to delete by ID (stored in status) even after a failed rename, but that is iffy. We can come back to that at a later time.

This is also adding integration tests to validate the desired behavior regarding default policies.