Validate scan watcher errors
This is aims to add granularity on the scan watcher error handling.
For example: if the scan timesout, was the CO installed with vesion >= 1.6? If not we add a more descriptive message to the report.

It also adds a check to make sure we do not start a new scan watcher if an old scan finishes. In other words, if scan finishes but we already handled a newer scan result, then we ignore this result.
This happens duo CO reporting scan updates with different timestamps:

- New scan is executed
- CO sends an update with timestamp X
- ScanWatcher starts for the timestamp X
- CO sends an update with timestamp Y
- ScanWatcher starts for the timestamp Y
- CO sends the CheckResults with time Y
- ScanWatcher for timestamp Y ends successfully
- ACS generates the report
- The ScanWatcher for timestamp X times out
- ACS ignores this error because we already have a scan reported with timestamp > X