ROX-25492, ROX-25493: Configure Policy-as-code in operator (#12799)
### CRD Installation in Operator

For operator-based installs of ACS, CRDs are currently installed by OLM.  While the `SecurityPolicy` CRD is unique in that it is actually used by the StackRox application, OLM is still the most intuitive method of installation.

This approach takes two steps:

1. Copy the CRD from config-controller over to the operator bundle
2. Add an ownership reference to the `SecurityPolicy` CRD in the ACS operator CSV

An alternative approach was considered where the `operator-sdk` CLI was "tricked" into generating the `SecurityPolicy` CRD, but this required copying over the go structs for the `SecurityPolicy` CRD, which seemed like much more work overall.

### `configAsCode` toggle in Helm Chart

This adds a Helm chart parameter to toggle config-as-code.  Defaults to `true`.

The name `configAsCode` is used instead of `policyAsCode` to make the API future-proof in case we add other config-related CRDs down the road.

Example:

```
configAsCode:
  enabled: false
```

### Make configAsCode configurable in operator

Exposing the above Helm chart toggle `configAsCode.enabled` in the `Central` CRD, defaulting to `Enabled`.

### Apply CRD in deploy script

Manifest installation will not include the CRD, so we are adding it "by hand" in the deploy script before either Helm or manifest installation methods have been executed.