This adds additonal checks to avoid watching unnecessary scans. So, if we received a scan that is not handled by ACS (does not have a ScanConfiguration associated), then we can ignore it because without a ScanConfiguration associated we cannot report on it.
It also changes the logic of failing the handling a ScanResults if the scan was already handled by the ScanConfigWatcher.
This was a problem because CO might send updates for the same scan execution with different timestamps and the old code will fail the ScanConfigWatcher if we received the second ScanResults.
The logic is changed in a way that now if the ScanResult is newer we override the old one and if it's older then we simply ignore it.