chore(tests): disable flaky network flow test
The test in question was reenabled when some additional logic was added
to prevent port 0 from being chosen, however, this is a deeper UDP issue
that we are not yet able to fully solve. Until we have time to look into
it, there's no point on having this flaky test around.