Instruct Renovate to create single PR per branch instead of multiple
It becomes burdensome to deal with the separate ones which get cut
around the same time. E.g.
https://github.com/stackrox/stackrox/pull/13479
https://github.com/stackrox/stackrox/pull/13478
https://github.com/stackrox/stackrox/pull/13477

What for? I don't see a valid reason.
I suggest to combine all such PRs in one for now until we have a case
for separation.

Used answers here as inspiration
https://stackoverflow.com/questions/66471226/renovate-combine-all-updates-to-one-branch-pr

I don't know yet if this will work. We need to merge, wait and see.