Switch `scan-images-with-roxctl` to work with quay.io/stackrox-io
Since the idea of this overall change is to disable GHA builds into
quay.io/rhacs-eng/ for releases, we can simply switch to scan
GHA-built images in quay.io/stackrox-io/ which should be there
(at least we don't have plans to migrate them to Konflux at this
point).

Some nerdy extra context:

In https://github.com/stackrox/stackrox/pull/13694, the similar
change for `.github/workflows/build.yaml`, I could not do the same
thing because of a situation with the `stackrox-operator` image: the
one is only built into quay.io/rhacs-eng and not built into
quay.io/stackrox-io. Therefore I had to introduce a matrix there for
the `scan-images-with-roxctl` job.

I was musing whether I should introduce the matrix here, in
`.github/workflows/scanner-build.yaml`, as well for this
`scan-images-with-roxctl` job. The benefit is consistency of
workflows. The downside is more complexity to this workflow.